+++ x/auth/migrations/v5/migrate.go
@@ -7,13 +7,12 @@ import (
 
 	"cosmossdk.io/collections"
 	storetypes "cosmossdk.io/core/store"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 )
-
-var LegacyGlobalAccountNumberKey = []byte("globalAccountNumber")
 
 func Migrate(ctx context.Context, storeService storetypes.KVStoreService, sequence collections.Sequence) error {
 	store := storeService.OpenKVStore(ctx)
-	b, err := store.Get(LegacyGlobalAccountNumberKey)
+	b, err := store.Get(authtypes.LegacyGlobalAccountNumberKey)
 	if err != nil {
 		return err
 	}
@@ -37,7 +36,7 @@ func Migrate(ctx context.Context, storeService storetypes.KVStoreService, sequen
 	}
 
 	// remove the value from the old prefix.
-	err = store.Delete(LegacyGlobalAccountNumberKey)
+	err = store.Delete(authtypes.LegacyGlobalAccountNumberKey)
 	if err != nil {
 		return err
 	}