+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * (baseapp) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) `SetProtocolVersion` has been renamed to `SetAppVersion`. It now updates the consensus params in baseapp's `ParamStore`.
 * (baseapp) [#1](https://github.com/01builders/cosmos-sdk/pull/1) Re-add query router for custom abci queries.
 * (testing) [#1](https://github.com/01builders/cosmos-sdk/pull/1) Add `MakeTestEncodingConfigWithOpts` to pass `testutil.CodecOptions` when creating `TestEncodingConfig`.
+* (x/auth) [#7](https://github.com/01builders/cosmos-sdk/pull/7) Support legacy global AccountNumber.
 
 ### Improvements